deps: update V8 to 8.6.395
PR-URL: https://github.com/nodejs/node/pull/35415

+++ deps/v8/src/torque/instructions.h
@@ -166,11 +166,11 @@ inline std::ostream& operator<<(std::ostream& stream,
                     << loc.GetParameterIndex() << ")";
     case DefinitionLocation::Kind::kPhi:
       return stream << "DefinitionLocation::Phi(" << std::hex
-                    << (uint64_t)loc.GetPhiBlock() << std::dec << ", "
+                    << loc.GetPhiBlock() << std::dec << ", "
                     << loc.GetPhiIndex() << ")";
     case DefinitionLocation::Kind::kInstruction:
       return stream << "DefinitionLocation::Instruction(" << std::hex
-                    << (uint64_t)loc.GetInstruction() << std::dec << ", "
+                    << loc.GetInstruction() << std::dec << ", "
                     << loc.GetInstructionIndex() << ")";
   }
 }
